added copy config and changed log/model structure
