+++ README.md
@@ -5,3 +5,6 @@
 - Install Python 3.10
 - Run ``pip install -r requirements.txt``
 - Run files with the green Play button in PyCharm
+
+#### Tensorboard
+- Window: promt cmd from project folder and run ``tensorboard --logdir logs``
