+++ src/common/utils.py
@@ -4,13 +4,13 @@ import shutil
 
 
 def get_config():
-    with open('common/config.yml') as f:
+    with open('config/configuration.yml') as f:
         data = yaml.load(f, Loader=SafeLoader)
 
         return data
 
 def copy_config(destination_dir):
-    shutil.copy(src='common/config.yml', dst=destination_dir)
+    shutil.copy(src='config/configuration.yml', dst=destination_dir)
 
 
 if __name__ == "__main__":
